nintendoapi/requirements.txt
grpcio>=1.66.0
grpcio-tools>=1.66.0
protobuf>=5.0.0

# MongoDB async client
motor>=3.5.0

# Validación de modelos de dominio
pydantic>=2.7.0
email-validator>=2.0.0.post2

# Utilidades varias
python-dotenv>=1.0.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
